requirements.txt
aiogram==3.0.0b7
SQLAlchemy==1.4
asyncpg==0.27.0
python-dotenv==1.0.0